Rename AuthenticateHeaders utility to AuthorizationHeaders
