+++ community/server/src/main/java/org/neo4j/server/rest/dbms/AuthorizationFilter.java
@@ -197,7 +197,7 @@ public class AuthorizationFilter implements Filter
         }
         else
         {
-            return AuthenticateHeaders.extractToken( value );
+            return AuthorizationHeaders.extractToken( value );
         }
     }
